+++ app/src/main/java/bav/astrobirthday/ui/settings/BirthdayUpdateWorker.kt
@@ -1,0 +1,26 @@
+package bav.astrobirthday.ui.settings
+
+import android.content.Context
+import androidx.work.CoroutineWorker
+import androidx.work.WorkerParameters
+import org.koin.core.component.KoinApiExtension
+import org.koin.core.component.KoinComponent
+import org.koin.core.component.inject
+
+@KoinApiExtension
+class BirthdayUpdateWorker(
+    context: Context,
+    params: WorkerParameters
+) : CoroutineWorker(context, params), KoinComponent {
+
+    private val syncPlanetsInfo: SyncPlanetsInfo by inject()
+
+    override suspend fun doWork(): Result {
+        syncPlanetsInfo.sync()
+        return Result.success()
+    }
+
+    companion object {
+        const val UNIQUE_WORK_NAME = "bav.astrobirthday.birthdays_update"
+    }
+}